re routed game rotues and user routes to different paths
