+++ server/routes/api/index.js
@@ -1,6 +1,8 @@
 const router = require('express').Router();
 const userRoutes = require('./userRoutes');
+const gameRoutes = require('./gameRoutes');
 
 router.use('/user', userRoutes);
+router.use('/game', gameRoutes);
 
 module.exports = router;
