+++ server/utils/auth.js
@@ -4,6 +4,28 @@ const secret = 'mysecretssshhhhhhh';
 const expiration = '4h';
 
 module.exports = {
+  authMiddleware: function (req, res, next) {
+    let token = req.query.token || req.headers.authorization;
+
+    if(req.headers.authorization) {
+      token = token.split(' ').pop().trim();
+    }
+
+    if(!token) {
+      return res.status(400).json({ message: 'You have no token!' })
+    }
+
+    try{
+      const {data} = jwt.verify(token, secret, {maxAge: expiration})
+      req.user = data;
+    } catch {
+      console.log('Invalid token');
+      return res.status(400).json({ message: 'Invalid token!'})
+    }
+
+    next();
+  },
+
   signToken: function ({ email, username, _id }) {
     const payload = { email, username, _id };
     return jwt.sign({ data: payload }, secret, { expiresIn: expiration });
